backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.26.0
aiomysql>=0.2.0
cryptography>=42.0.0
